requirements.txt
alembic==1.4.2
attrs==20.1.0
cffi==1.14.2
click==7.1.2
cryptography==3.0
ecdsa==0.14.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==22.0.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
Werkzeug==1.0.1
python-dotenv
